+++ test/unit/events/orders/orders.command.spec.ts
@@ -6,17 +6,17 @@ import {
 	OrderFulfilledCommand,
 	OrderPaidCommand,
 	OrderRefundedCommand
-} from "../../../src/substrate/orders";
-import { OrderCancelledHandler } from "../../../src/substrate/orders/commands/order-cancelled/order-cancelled.handler";
-import { OrderCreatedHandler } from "../../../src/substrate/orders/commands/order-created/order-created.handler";
-import { OrderFailedHandler } from "../../../src/substrate/orders/commands/order-failed/order-failed.handler";
-import { OrderFulfilledHandler } from "../../../src/substrate/orders/commands/order-fulfilled/order-fulfilled.handler";
-import { OrderPaidHandler } from "../../../src/substrate/orders/commands/order-paid/order-paid.handler";
-import { OrderRefundedHandler } from "../../../src/substrate/orders/commands/order-refunded/order-refunded.handler";
-import { CommonModule } from "../../../src/common/common.module";
-import { OrderStatus } from "../../../src/substrate/orders/models/order-status";
-import { Currency } from "../../../src/substrate/orders/models/currency";
-import { BlockMetaData } from "../../../src/substrate/models/blockMetaData";
+} from "../../../../src/substrate/events/orders";
+import { OrderCancelledHandler } from "../../../../src/substrate/events/orders/commands/order-cancelled/order-cancelled.handler";
+import { OrderCreatedHandler } from "../../../../src/substrate/events/orders/commands/order-created/order-created.handler";
+import { OrderFailedHandler } from "../../../../src/substrate/events/orders/commands/order-failed/order-failed.handler";
+import { OrderFulfilledHandler } from "../../../../src/substrate/events/orders/commands/order-fulfilled/order-fulfilled.handler";
+import { OrderPaidHandler } from "../../../../src/substrate/events/orders/commands/order-paid/order-paid.handler";
+import { OrderRefundedHandler } from "../../../../src/substrate/events/orders/commands/order-refunded/order-refunded.handler";
+import { CommonModule } from "../../../../src/common/common.module";
+import { OrderStatus } from "../../../../src/substrate/events/orders/models/order-status";
+import { Currency } from "../../../../src/substrate/events/orders/models/currency";
+import { BlockMetaData } from "../../../../src/substrate/models/blockMetaData";
 import {
 	ElasticsearchModule,
 	ElasticsearchService
@@ -28,8 +28,15 @@ import {
 	Test,
 	TestingModule
 } from "@nestjs/testing";
-import { SubstrateController, SubstrateService } from "../../../src/substrate/substrate.handler";
-import { CommandBusProvider, ElasticSearchServiceProvider, substrateServiceProvider } from "../mock";
+import { 
+	SubstrateController, 
+	SubstrateService 
+} from "../../../../src/substrate/substrate.handler";
+import { 
+	CommandBusProvider, 
+	ElasticSearchServiceProvider, 
+	substrateServiceProvider 
+} from "../../mock";
 
 let orderCancelledHandler: OrderCancelledHandler;
 let orderCreatedHandler: OrderCreatedHandler;
