+++ test/unit/events/service-request/service-request.command.spec.ts
@@ -13,18 +13,30 @@ import {
 import {
 	SubstrateController,
 	SubstrateService
-} from "../../../src/substrate/substrate.handler";
-import { CommonModule } from "../../../src/common/common.module";
-import { ClaimedServiceRequestHandler } from "../../../src/substrate/service-request/commands/claimed-service-request/claimed-service-request.handler";
-import { CreateServiceRequestHandler } from "../../../src/substrate/service-request/commands/create-service-request/create-service-request.handler";
-import { FinalizedServiceRequestHandler } from "../../../src/substrate/service-request/commands/finalized-service-request/finalized-service-request.handler";
-import { ProcessedServiceRequestHandler } from "../../../src/substrate/service-request/commands/processed-service-request/processed-service-request.handler";
-import { UnstakedServiceRequestHandler } from "../../../src/substrate/service-request/commands/unstaked-service-request/unstaked-service-request.handler";
-import { UnstakedWaitingServiceRequestHandler } from "../../../src/substrate/service-request/commands/unstakedwaiting-service-request/unstakedwaiting-service-request.handler";
-import { ClaimedServiceRequestCommand, CreateServiceRequestCommand, FinalizedServiceRequestCommand, ProcessedServiceRequestCommand, RequestServiceCommandHandlers, UnstakedServiceRequestCommand, UnstakedWaitingServiceRequestCommand } from "../../../src/substrate/service-request";
-import { RequestStatus } from "../../../src/substrate/service-request/models/requestStatus";
-import { BlockMetaData } from "../../../src/substrate/models/blockMetaData";
-import { CommandBusProvider, ElasticSearchServiceProvider, substrateServiceProvider } from "../mock";
+} from "../../../../src/substrate/substrate.handler";
+import { CommonModule } from "../../../../src/common/common.module";
+import { ClaimedServiceRequestHandler } from "../../../../src/substrate/events/service-request/commands/claimed-service-request/claimed-service-request.handler";
+import { CreateServiceRequestHandler } from "../../../../src/substrate/events/service-request/commands/create-service-request/create-service-request.handler";
+import { FinalizedServiceRequestHandler } from "../../../../src/substrate/events/service-request/commands/finalized-service-request/finalized-service-request.handler";
+import { ProcessedServiceRequestHandler } from "../../../../src/substrate/events/service-request/commands/processed-service-request/processed-service-request.handler";
+import { UnstakedServiceRequestHandler } from "../../../../src/substrate/events/service-request/commands/unstaked-service-request/unstaked-service-request.handler";
+import { UnstakedWaitingServiceRequestHandler } from "../../../../src/substrate/events/service-request/commands/unstakedwaiting-service-request/unstakedwaiting-service-request.handler";
+import { 
+	ClaimedServiceRequestCommand, 
+	CreateServiceRequestCommand, 
+	FinalizedServiceRequestCommand, 
+	ProcessedServiceRequestCommand, 
+	RequestServiceCommandHandlers, 
+	UnstakedServiceRequestCommand, 
+	UnstakedWaitingServiceRequestCommand 
+} from "../../../../src/substrate/events/service-request";
+import { RequestStatus } from "../../../../src/substrate/events/service-request/models/requestStatus";
+import { BlockMetaData } from "../../../../src/substrate/models/blockMetaData";
+import { 
+	CommandBusProvider, 
+	ElasticSearchServiceProvider, 
+	substrateServiceProvider 
+} from "../../mock";
 
 let claimedServiceRequestHandler: ClaimedServiceRequestHandler;
 let createServiceRequestHandler: CreateServiceRequestHandler;
