+++ test/unit/events/services/services.command.spec.ts
@@ -12,18 +12,22 @@ import {
   ServiceCreatedCommand,
   ServiceDeletedCommand,
   ServiceUpdatedCommand
-} from "../../../src/substrate/services";
-import { CommonModule } from "../../../src/common/common.module";
-import { BlockMetaData } from "../../../src/substrate/models/blockMetaData";
-import { ServiceFlow } from "../../../src/substrate/models/service-flow";
-import { ServiceCreatedHandler } from "../../../src/substrate/services/commands/service-created/service-created.handler";
-import { ServiceDeletedHandler } from "../../../src/substrate/services/commands/service-deleted/service-deleted.handler";
-import { ServiceUpdatedHandler } from "../../../src/substrate/services/commands/service-updated/service-updated.handler";
+} from "../../../../src/substrate/events/services";
+import { CommonModule } from "../../../../src/common/common.module";
+import { BlockMetaData } from "../../../../src/substrate/models/blockMetaData";
+import { ServiceFlow } from "../../../../src/substrate/models/service-flow";
+import { ServiceCreatedHandler } from "../../../../src/substrate/events/services/commands/service-created/service-created.handler";
+import { ServiceDeletedHandler } from "../../../../src/substrate/events/services/commands/service-deleted/service-deleted.handler";
+import { ServiceUpdatedHandler } from "../../../../src/substrate/events/services/commands/service-updated/service-updated.handler";
 import {
 	SubstrateController,
 	SubstrateService
-} from "../../../src/substrate/substrate.handler";
-import { CommandBusProvider, ElasticSearchServiceProvider, substrateServiceProvider } from "../mock";
+} from "../../../../src/substrate/substrate.handler";
+import { 
+	CommandBusProvider, 
+	ElasticSearchServiceProvider, 
+	substrateServiceProvider 
+} from "../../mock";
 
 let serviceCreatedHandler: ServiceCreatedHandler;
 let serviceDeletedHandler: ServiceDeletedHandler;
